# docs/openapi/swagger.yaml
definitions:
  api_rest_admin.SetUseInTrainingRequest:
    properties:
      use_in_training:
        type: boolean
    type: object
  api_rest_admin.StrudelAdminResponse:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      use_in_training:
        type: boolean
      user_id:
        type: string
    type: object
  api_rest_agent.DocReference:
    properties:
      page_name:
        type: string
      section_title:
        type: string
      url:
        type: string
    type: object
  api_rest_agent.GenerateRequest:
    properties:
      conversation_history:
        items:
          $ref: '#/definitions/api_rest_agent.Message'
        type: array
      editor_state:
        type: string
      forked_from_id:
        description: 'optional: for blocking AI on restricted forks'
        type: string
      provider:
        description: '"anthropic" or "openai"'
        type: string
      provider_api_key:
        description: BYOK key
        type: string
      session_id:
        description: 'optional: for paste lock validation'
        type: string
      strudel_id:
        description: 'optional: for persisting conversation'
        type: string
      user_query:
        type: string
    required:
    - user_query
    type: object
  api_rest_agent.GenerateResponse:
    properties:
      clarifying_questions:
        items:
          type: string
        type: array
      code:
        type: string
      doc_references:
        items:
          $ref: '#/definitions/api_rest_agent.DocReference'
        type: array
      docs_retrieved:
        type: integer
      examples_retrieved:
        type: integer
      is_actionable:
        type: boolean
      is_code_response:
        type: boolean
      model:
        type: string
      strudel_references:
        items:
          $ref: '#/definitions/api_rest_agent.StrudelReference'
        type: array
    type: object
  api_rest_agent.Message:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  api_rest_agent.StrudelReference:
    properties:
      author_name:
        type: string
      id:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
...
  codeberg_org_algopatterns_server_internal_errors.ErrorResponse:
    properties:
      details:
        description: optional details (sanitized in production)
        type: string
      error:
        description: error code (e.g., "unauthorized", "not_found")
        type: string
      message:
        description: user-friendly message
        type: string
      request_id:
        description: echoed from request for correlation
        type: string
    type: object
host: algopatterns.dev
info:
  contact:
    name: API Support
    url: https://codeberg.org/algopatterns/server
  description: |-
    AI-powered Strudel code generation and collaborative live coding platform

    Features:
    - AI-powered Strudel code generation from natural language
    - Real-time collaborative editing via WebSockets
    - OAuth authentication (Google, GitHub)
    - Anonymous session support
    - Save and share Strudel patterns
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Algopatterns API
  version: "1.0"
paths:
  /api/v1/admin/strudels/{id}:
    get:
      description: Admin-only endpoint to get any strudel regardless of ownership
      parameters:
      - description: Strudel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_admin.StrudelAdminResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - AdminKeyAuth: []
      summary: Get any strudel by ID (admin)
      tags:
      - admin
  /api/v1/admin/strudels/{id}/use-in-training:
    put:
      consumes:
      - application/json
      description: Admin-only endpoint to mark a strudel for use in training data
      parameters:
      - description: Strudel ID
        in: path
        name: id
        required: true
        type: string
      - description: Use in training data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_rest_admin.SetUseInTrainingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_admin.StrudelAdminResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - AdminKeyAuth: []
      summary: Set use_in_training flag on a strudel
      tags:
      - admin
  /api/v1/agent/generate:
    post:
      consumes:
      - application/json
      description: Generate Strudel code using AI with optional BYOK support
      parameters:
      - description: Generation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_rest_agent.GenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_agent.GenerateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      summary: Generate code with AI
      tags:
      - agent
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token for authenticated requests. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"